+++ src/main/java/de/srendi/advancedperipherals/common/addons/computercraft/peripheral/PlayerDetectorPeripheral.java
@@ -61,7 +61,7 @@ public class PlayerDetectorPeripheral extends BasePeripheral<IPeripheralOwner> {
 
     @LuaFunction(mainThread = true)
     public final MethodResult getPlayersInCoords(Map<?, ?> firstCoord, Map<?, ?> secondCoord) throws LuaException {
-        if (owner.getOwner() == null)
+        if (isOnPocket() && owner.getOwner() == null)
             return MethodResult.of(null, "NOT_ATTACHED_TO_PLAYER");
 
         List<String> playersName = new ArrayList<>();
@@ -80,7 +80,7 @@ public class PlayerDetectorPeripheral extends BasePeripheral<IPeripheralOwner> {
 
     @LuaFunction(mainThread = true)
     public final MethodResult getPlayersInCubic(int x, int y, int z) {
-        if (owner.getOwner() == null)
+        if (isOnPocket() && owner.getOwner() == null)
             return MethodResult.of(null, "NOT_ATTACHED_TO_PLAYER");
 
         List<String> playersName = new ArrayList<>();
@@ -97,7 +97,7 @@ public class PlayerDetectorPeripheral extends BasePeripheral<IPeripheralOwner> {
 
     @LuaFunction(mainThread = true)
     public final MethodResult getPlayersInRange(int range) {
-        if (owner.getOwner() == null)
+        if (isOnPocket() && owner.getOwner() == null)
             return MethodResult.of(null, "NOT_ATTACHED_TO_PLAYER");
 
         List<String> playersName = new ArrayList<>();
@@ -114,7 +114,7 @@ public class PlayerDetectorPeripheral extends BasePeripheral<IPeripheralOwner> {
 
     @LuaFunction(mainThread = true)
     public final MethodResult isPlayersInCoords(Map<?, ?> firstCoord, Map<?, ?> secondCoord) throws LuaException {
-        if (owner.getOwner() == null)
+        if (isOnPocket() && owner.getOwner() == null)
             return MethodResult.of(false, "NOT_ATTACHED_TO_PLAYER");
 
         if (getPlayers().isEmpty())
@@ -133,7 +133,7 @@ public class PlayerDetectorPeripheral extends BasePeripheral<IPeripheralOwner> {
 
     @LuaFunction(mainThread = true)
     public final MethodResult isPlayersInCubic(int x, int y, int z) {
-        if (owner.getOwner() == null)
+        if (isOnPocket() && owner.getOwner() == null)
             return MethodResult.of(false, "NOT_ATTACHED_TO_PLAYER");
 
         if (getPlayers().isEmpty())
@@ -150,7 +150,7 @@ public class PlayerDetectorPeripheral extends BasePeripheral<IPeripheralOwner> {
 
     @LuaFunction(mainThread = true)
     public final MethodResult isPlayersInRange(int range) {
-        if (owner.getOwner() == null)
+        if (isOnPocket() && owner.getOwner() == null)
             return MethodResult.of(false, "NOT_ATTACHED_TO_PLAYER");
 
         if (getPlayers().isEmpty())
@@ -167,7 +167,7 @@ public class PlayerDetectorPeripheral extends BasePeripheral<IPeripheralOwner> {
 
     @LuaFunction(mainThread = true)
     public final MethodResult isPlayerInCoords(Map<?, ?> firstCoord, Map<?, ?> secondCoord, String username) throws LuaException {
-        if (owner.getOwner() == null)
+        if (isOnPocket() && owner.getOwner() == null)
             return MethodResult.of(false, "NOT_ATTACHED_TO_PLAYER");
 
         BlockPos firstPos = LuaConverter.convertToBlockPos(firstCoord);
@@ -186,7 +186,7 @@ public class PlayerDetectorPeripheral extends BasePeripheral<IPeripheralOwner> {
 
     @LuaFunction(mainThread = true)
     public final MethodResult isPlayerInCubic(int x, int y, int z, String username) {
-        if (owner.getOwner() == null)
+        if (isOnPocket() && owner.getOwner() == null)
             return MethodResult.of(false, "NOT_ATTACHED_TO_PLAYER");
 
         ResourceKey<Level> dimension = getLevel().dimension();
@@ -204,7 +204,7 @@ public class PlayerDetectorPeripheral extends BasePeripheral<IPeripheralOwner> {
 
     @LuaFunction(mainThread = true)
     public final MethodResult isPlayerInRange(int range, String username) {
-        if (owner.getOwner() == null)
+        if (isOnPocket() && owner.getOwner() == null)
             return MethodResult.of(false, "NOT_ATTACHED_TO_PLAYER");
 
         ResourceKey<Level> dimension = getLevel().dimension();
@@ -225,7 +225,7 @@ public class PlayerDetectorPeripheral extends BasePeripheral<IPeripheralOwner> {
         if (!APConfig.PERIPHERALS_CONFIG.playerSpy.get())
             throw new LuaException("This function is disabled in the config. Activate it or ask an admin if he can activate it.");
 
-        if (owner.getOwner() == null)
+        if (isOnPocket() && owner.getOwner() == null)
             return MethodResult.of(null, "NOT_ATTACHED_TO_PLAYER");
 
         ResourceKey<Level> dimension = getLevel().dimension();